requirement.txt
coverage==6.3.3
Django==3.2
djangorestframework==3.13.1
